requirements.txt
python-can
python-telegram-bot
aiogram